Split Printer to its own project, referenced by both LinqToGraphql and GraphQLSchemaToCSharp
