+++ net7.0/Telia.GraphQLPrinter/Telia.GraphQLPrinter/Telia.GraphQLPrinter.csproj
@@ -1,17 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>disable</Nullable>
+    <Nullable>enable</Nullable>
     <Configurations>Release</Configurations>
   </PropertyGroup>
+
   <ItemGroup>
     <PackageReference Include="GraphQL-Parser" Version="8.1.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
-    <PackageReference Include="System.Text.Json" Version="6.0.7" />
     <PackageReference Include="SystemLibrary.Common.Net" Version="6.6.1.4" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Telia.GraphQLPrinter\Telia.GraphQLPrinter\Telia.GraphQLPrinter.csproj" />
-  </ItemGroup>
+
 </Project>
