+++ net7.0/Telia.GraphQLSchemaToCSharp.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.GraphQLSchemaToCSharp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.GraphQLSchemaToCSharp.Tests", "Telia.GraphQLSchemaToCSharp.Tests\Telia.GraphQLSchemaToCSharp.Tests.csproj", "{903FC7F1-8D61-4DBE-AFB3-42501DF08820}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.GraphQLPrinter", "Telia.GraphQLPrinter\Telia.GraphQLPrinter\Telia.GraphQLPrinter.csproj", "{E5567D33-EB93-4089-898B-234414AE08F2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Release|Any CPU = Release|Any CPU
@@ -16,6 +18,8 @@ Global
 		{794782F0-6869-4277-8717-C5B8F9919D3C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{903FC7F1-8D61-4DBE-AFB3-42501DF08820}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{903FC7F1-8D61-4DBE-AFB3-42501DF08820}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5567D33-EB93-4089-898B-234414AE08F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5567D33-EB93-4089-898B-234414AE08F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
