+++ net7.0/Telia.GraphQLSchemaToCSharp/SchemaConverter/DefinitionHandlers/InputDefinitionHandler.cs
@@ -1,9 +1,9 @@
-﻿using GraphQLParser;
-using GraphQLParser.AST;
+﻿using GraphQLParser.AST;
 
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
+using Telia.GraphQLPrinter;
 using Telia.GraphQLSchemaToCSharp.Config;
 
 namespace Telia.GraphQLSchemaToCSharp.DefinitionHandlers;
