+++ net7.0/Telia.LinqToGraphQLToModel.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.LinqToGraphQLToModel.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.LinqToGraphQLToModel.Tests", "Telia.LinqToGraphQLToModel.Tests\Telia.LinqToGraphQLToModel.Tests.csproj", "{ADA7FD63-47AC-45AB-9767-863D916B5EAF}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Telia.GraphQLPrinter", "Telia.GraphQLPrinter\Telia.GraphQLPrinter\Telia.GraphQLPrinter.csproj", "{A1D2CC2F-BE22-404E-A742-F39E7598C4F5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Release|Any CPU = Release|Any CPU
@@ -20,6 +22,8 @@ Global
 		{B7ADBA87-4E22-4F1B-BE5A-F3ECAA0C8437}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ADA7FD63-47AC-45AB-9767-863D916B5EAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ADA7FD63-47AC-45AB-9767-863D916B5EAF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1D2CC2F-BE22-404E-A742-F39E7598C4F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1D2CC2F-BE22-404E-A742-F39E7598C4F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
