+++ net7.0/Telia.LinqToGraphQLToModel/SelectionChainConverter.cs
@@ -1,5 +1,6 @@
 ﻿using GraphQLParser.AST;
 
+using Telia.GraphQLPrinter.Models;
 using Telia.LinqToGraphQLToModel.Models;
 
 namespace Telia.LinqToGraphQLToModel;
@@ -133,19 +134,4 @@ internal class SelectionChainConverter
 
         return variable;
     }
-}
-
-public class GraphQLFieldNode : ASTNode, IHasSelectionSetNode, IHasArgumentsNode, IHasDirectivesNode, INamedNode
-{
-    public GraphQLName Alias { get; set; }
-
-    public override ASTNodeKind Kind => ASTNodeKind.Field;
-
-    public GraphQLName Name { get; set; }
-
-    public GraphQLSelectionSet SelectionSet { get; set; }
-
-    public GraphQLDirectives Directives { get; set; }
-
-    public GraphQLArguments Arguments { get; set; }
 }
